refactor(ime): merge TextInputManager into TrimeInputMethodService
TextInputManager class, which allowed us to use Kotlin to assist TrimeInputMethodService when there was a lot of Java code in the project, is now complete
